login  con errores

+++ src/main/java/com/example/DoloresAleman_Final/service/DomicilioService.java
@@ -2,6 +2,7 @@ package com.example.DoloresAleman_Final.service;
 
 import com.example.DoloresAleman_Final.Model.DomicilioDTO;
 import com.example.DoloresAleman_Final.controller.OdontologoController;
+import com.example.DoloresAleman_Final.exceptions.ResourceNotFoundException;
 import com.example.DoloresAleman_Final.persistence.entity.Domicilio;
 import com.example.DoloresAleman_Final.persistence.repository.IDomicilioRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -37,7 +38,13 @@ public class DomicilioService {
         repository.save(domicilio);
     }*/
 
-    public void eliminar(Long id){
+    /*public void eliminar(Long id){
+        repository.deleteById(id);
+    }*/
+    //mismo metodo pero con excepciones
+    public void eliminar(Long id) throws ResourceNotFoundException {
+        if (buscarPorId(id) == null)
+            throw new ResourceNotFoundException("no existe un domicilio con id: " + id);
         repository.deleteById(id);
     }
